removed layout specific properties from network changed float->double moved all types declarations to types.h

+++ flowlayout/src/functions.cpp
@@ -17,11 +17,11 @@ Point operator-(const Point p1, const Point p2){
    return p;   
 }
 
-float operator*(const Point p1, const Point p2){
+double operator*(const Point p1, const Point p2){
    //vector product of two vectors
    return p1.x*p2.y-p1.y*p2.x;
 }
-Point operator*(const Point p1, const float scalar){
+Point operator*(const Point p1, const double scalar){
    //scale vector
    Point p;
    p.x=p1.x*scalar;
@@ -29,43 +29,43 @@ Point operator*(const Point p1, const float scalar){
    return p;
 }
 Point unit(const Point p1){
-   float len=norm(p1);
+   double len=norm(p1);
    Point p=p1;
    return p*(1/len);
 }
 
-Point to_left(const Point p0, const float beta){
+Point to_left(const Point p0, const double beta){
    //rotating the vector "p0" "beta" degrees to the left.
-   float alpha=angle(p0), d=norm(p0);
+   double alpha=angle(p0), d=norm(p0);
    Point p;
    p.x=d*cos(alpha+beta);
    p.y=d*sin(alpha+beta);
    return p;
 }
 
-float norm(const Point p){
+double norm(const Point p){
    //norm (length) of a vector.
    return sqrt(p.x*p.x+p.y*p.y);
 }
 
-float angle(const Point p){
+double angle(const Point p){
    // angle of a vector (w.r.t +x axis), within range [0.5PI, 1.5PI].
    return atan2(p.y,p.x);
 /*   if(p.x==0){
       if(p.y>=0)return 0.5*PI;
       else return -0.5*PI;
    }
-   float alpha=atan(p.y/p.x);
+   double alpha=atan(p.y/p.x);
    if(p.x<0)alpha+=PI; //make sure it fit in range [0.5PI, 1.5PI].
    return alpha;*/
 }
 
-float dist(const Point p1, const Point p2){
+double dist(const Point p1, const Point p2){
    //the distance between point p1 and p2.
    return norm(p1-p2);
 }
 
-float lim(float beta){
+double lim(double beta){
    //make an angle to fit in range [-PI, PI].
    while (beta<-PI) {
       beta+=(2*PI);
